freezed requirements
