+++ .gitignore
@@ -1,4 +1,5 @@
 secrate_code.txt
+__VC__.txt
 # Created by https://www.gitignore.io/api/linux,macos,python,django,windows,pycharm,visualstudiocode
 # Edit at https://www.gitignore.io/?templates=linux,macos,python,django,windows,pycharm,visualstudiocode
