+++ important links.txt
@@ -1,2 +1,21 @@
 https://wsvincent.com/django-rest-framework-user-authentication-tutorial/
 https://www.django-rest-framework.org/topics/html-and-forms/
+https://testdriven.io/blog/django-custom-user-model/
+https://wsvincent.com/django-custom-user-model-tutorial/
+https://dev.to/callmetarush/the-django-rest-custom-user-model-and-authentication-5go9
+https://medium.com/@apogiatzis/create-a-restful-api-with-users-and-jwt-authentication-using-django-1-11-drf-part-1-288268602bb7
+https://simpleisbetterthancomplex.com/tutorial/2016/07/22/how-to-extend-django-user-model.html
+http://fabcoding.com/building-an-api-with-django-rest-framework-2-user-authentication/
+https://www.django-rest-framework.org/tutorial/4-authentication-and-permissions/
+https://medium.com/backticks-tildes/lets-build-an-api-with-django-rest-framework-part-2-cfb87e2c8a6c
+https://iheanyi.com/journal/user-registration-authentication-with-django-django-rest-framework-react-and-redux/
+https://pypi.org/project/django-rest-registration/
+https://django-rest-registration.readthedocs.io/en/latest/
+https://github.com/beingaskar/django-rest-framework-user-registration
+https://sunscrapers.com/blog/django-rest-framework-login-and-authentication/
+https://github.com/jpadilla/django-rest-framework-jwt
+https://jpadilla.github.io/django-rest-framework-jwt/
+https://simpleisbetterthancomplex.com/tutorial/2018/12/19/how-to-use-jwt-authentication-with-django-rest-framework.html
+https://www.techiediaries.com/django-rest-framework-jwt-tutorial/
+https://blog.joeymasip.com/how-to-build-a-rest-api-with-django-and-jwt-auth/
+https://thinkster.io/tutorials/django-json-api/authentication
